+++ Plugin~/WebRTCPlugin/GraphicsDevice/Vulkan/CMakeLists.txt
@@ -1,6 +1,8 @@
 target_sources(WebRTCLib
   PRIVATE
     ListOfVulkanFunctions.inl
+    LoadVulkanFunctions.cpp
+    LoadVulkanFunctions.h
     UnityVulkanInitCallback.cpp
     UnityVulkanInitCallback.h
     VulkanGraphicsDevice.cpp
