+++ Plugin~/WebRTCPlugin/GraphicsDevice/Vulkan/ListOfVulkanFunctions.inl
@@ -21,7 +21,6 @@ GLOBAL_VULKAN_FUNCTION(vkCreateInstance)
 
 INSTANCE_VULKAN_FUNCTION(vkEnumeratePhysicalDevices)
 INSTANCE_VULKAN_FUNCTION(vkEnumerateDeviceExtensionProperties)
-//INSTANCE_VULKAN_FUNCTION(vkGetPhysicalDeviceProperties)
 INSTANCE_VULKAN_FUNCTION(vkGetPhysicalDeviceQueueFamilyProperties)
 INSTANCE_VULKAN_FUNCTION(vkGetPhysicalDeviceMemoryProperties)
 INSTANCE_VULKAN_FUNCTION(vkGetPhysicalDeviceFormatProperties)
